Adds simplified project for SensorVehicle-main
This is a simplified copy of the SensorVehicle-main project from the main repo at:
https://github.com/BjAlvestad/BO19E-15_SensorVehicle

The projects in the solution that the students does not require access to for creating their control logics has been comiled to dll files and added as refrences.
This was done to make the program less overwhelming when first starting out.
XML files with code documentation is also included so that IntelliSense will still help with descriptions of the functions.

The full code, and newest experimental version is available on:
https://github.com/BjAlvestad/BO19E-15_SensorVehicle